# server/pkg/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.PlaidAccountsReconnectedRequest:
    properties:
      accounts:
        items:
          $ref: '#/definitions/plaid.PlaidLinkAccount'
        type: array
      linkSessionId:
        type: string
    required:
    - accounts
    - linkSessionId
    type: object
...
  handler.RecoverOnboardingOtpRequest:
    properties:
      type:
        type: string
    required:
    - type
    type: object
  handler.RecoverOnboardingOtpResponse:
    properties:
      maskedEmail:
        type: string
      maskedMobileNo:
        type: string
      otpExpiryDuration:
        type: integer
      otpId:
        type: string
    required:
    - maskedEmail
    - maskedMobileNo
    - otpExpiryDuration
    - otpId
    type: object
...
  plaid.PlaidLinkAccount:
    properties:
      id:
        type: string
      mask:
        type: string
      name:
        type: string
      subtype:
        enum:
        - checking
        - savings
        type: string
      type:
        type: string
      verificationStatus:
        type: string
    required:
    - id
    - subtype
    - type
    type: object
...
  request.DemographicUpdateSendOtpRequest:
    properties:
      type:
        type: string
    required:
    - type
    type: object
...
  response.BadRequestError:
    properties:
      error:
        type: string
      fieldName:
        type: string
    required:
    - error
    type: object
  response.BadRequestErrors:
    properties:
      errors:
        items:
          $ref: '#/definitions/response.BadRequestError'
        type: array
    required:
    - errors
    type: object
...
  response.DemographicUpdateSendOtpResponse:
    properties:
      maskedEmail:
        type: string
      maskedMobileNo:
        type: string
      otpExpiryDuration:
        type: integer
      otpId:
        type: string
    required:
    - maskedEmail
    - maskedMobileNo
    - otpExpiryDuration
    - otpId
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: string
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: DreamFi Middleware API
  version: "1.0"
paths:
...
  /account/customer/demographic-update/mobile:
    post:
      consumes:
      - application/json
      description: Generates and sends an OTP code for the current user before initiating
        a demographic update
      parameters:
      - description: Payload indicating OTP type
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/request.DemographicUpdateSendOtpRequest'
      - description: Bearer token for user authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Authorization:
              description: Bearer token for user authentication
              type: string
          schema:
            $ref: '#/definitions/response.DemographicUpdateSendOtpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: DemographicUpdateSendOtp
      tags:
      - Demographic Updates
...
  /account/plaid/accounts/reconnected:
    post:
      description: Registers that accounts were reconnected after Plaid Link was run
        in Update mode
      parameters:
      - description: PlaidAccountsReconnectedRequest
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.PlaidAccountsReconnectedRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Ok
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: PlaidAccountsReconnected
      tags:
      - plaid
...
  /recover-onboarding/send-otp:
    post:
      consumes:
      - application/json
      description: Sends an onboarding recovery OTP to the user
      parameters:
      - description: RecoverOnboardingOtpRequest
        in: body
        name: recoverOnboardingOtpRequest
        required: true
        schema:
          $ref: '#/definitions/handler.RecoverOnboardingOtpRequest'
      - description: Bearer token for user authentication
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: RecoverOnboardingOtpResponse
          headers:
            Authorization:
              description: Bearer token for user authentication
              type: string
          schema:
            $ref: '#/definitions/handler.RecoverOnboardingOtpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestErrors'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "412":
          description: Precondition Failed
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: RecoverOnboardingOTP
...
swagger: "2.0"